add handler, factory and classes managing next computer move depending on difficulty leve
