+++ src/ComputerMoveHandler.java
@@ -1,0 +1,3 @@
+public interface ComputerMoveHandler {
+    int[] getNextMove();
+}
